chore: Update development dependency group in poetry for compatibility
- Updated development dependency group for improved compatibility

A note about defining a dev dependencies group

The proper way to define a dev dependencies group since Poetry 1.2.0 is the following:

```toml
[tool.poetry.group.dev.dependencies]
pytest = "^6.0.0"
pytest-mock = "*"
```

This group notation is preferred since Poetry 1.2.0 and not usable in earlier versions. For backwards compatibility with older versions of Poetry, any dependency declared in the dev-dependencies section will automatically be added to the dev group. So the above and following notations are equivalent:

```toml
# Poetry pre-1.2.x style, understood by Poetry 1.0–1.2
[tool.poetry.dev-dependencies]
pytest = "^6.0.0"
pytest-mock = "*"
```

Poetry will slowly transition away from the dev-dependencies notation which will soon be deprecated, so it’s advised to migrate your existing development dependencies to the new group notation.